Use the path to hubconf.py to find models
While `torch.hub.load` uses predictable names when downloading from github,
the name referenced only worked when using the `master` branch of the
repo. Using a ref like `snakers4/silero_vad:v4.0` results in a
different directory name and the model files not being found.

Since the model files are stored in a path relative to hubconf.py, use
the path of hubconf.py to find the models instead of assuming the
directory `torch.hub.load` will extract the files to.